Add comments to all classes. Add english localization Move the logic for event processing from the main class InfiniteWaterMod to CommonProxy and utilize the @Mod.EventHandler annotation to handle events directly in the proxy class. Add constants to centrally manage strings and avoid hardcoding
