+++ src/main/resources/assets/infinitewater/lang/en_US.lang
@@ -1,0 +1,1 @@
+tile.infinite_water_block.name=Infinite Water Block
